requirements.txt
numpy
tqdm
matplotlib
tensorflow
psutil
grpcio